+++ resources/js/components/RoomTabRecordingsEditButton.vue
@@ -5,12 +5,12 @@
     severity="info"
     icon="fa-solid fa-edit"
     data-test="room-recordings-edit-button"
-    @click="showEditModal"
+    @click="showModal"
   />
 
   <!-- edit recording modal -->
   <Dialog
-    v-model:visible="showModal"
+    v-model:visible="modalVisible"
     modal
     :style="{ width: '500px' }"
     :breakpoints="{ '575px': '90vw' }"
@@ -41,7 +41,7 @@
           severity="secondary"
           :disabled="isLoadingAction"
           data-test="dialog-cancel-button"
-          @click="showModal = false"
+          @click="modalVisible = false"
         />
         <Button
           :label="$t('app.save')"
@@ -174,7 +174,7 @@ const settingsStore = useSettingsStore();
 const toast = useToast();
 const { t } = useI18n();
 
-const showModal = ref(false);
+const modalVisible = ref(false);
 const newDescription = ref(null);
 const newFormats = ref([]);
 const newAccess = ref(null);
@@ -196,12 +196,12 @@ const charactersLeftDescription = computed(() => {
 /**
  * show modal to edit recording
  */
-function showEditModal() {
+function showModal() {
   newDescription.value = props.description;
   newFormats.value = _.cloneDeep(props.formats);
   newAccess.value = props.access;
   formErrors.clear();
-  showModal.value = true;
+  modalVisible.value = true;
 }
 
 /**
@@ -224,7 +224,7 @@ function save() {
     })
     .then(() => {
       // operation successful, close modal and reload list
-      showModal.value = false;
+      modalVisible.value = false;
       emit("edited");
     })
     .catch((error) => {
@@ -233,7 +233,7 @@ function save() {
         // recording not found
         if (error.response.status === env.HTTP_NOT_FOUND) {
           toast.error(t("rooms.flash.recording_gone"));
-          showModal.value = false;
+          modalVisible.value = false;
           emit("notFound");
           return;
         }