+++ resources/js/components/RoomTabRecordingsViewButton.vue
@@ -5,12 +5,12 @@
     icon="fa-solid fa-eye"
     :disabled="props.disabled"
     data-test="room-recordings-view-button"
-    @click="showModal = true"
+    @click="modalVisible = true"
   />
 
   <!-- view recording modal -->
   <Dialog
-    v-model:visible="showModal"
+    v-model:visible="modalVisible"
     modal
     :style="{ width: '500px' }"
     :breakpoints="{ '575px': '90vw' }"
@@ -42,7 +42,7 @@
           icon="fa-solid fa-times"
           :disabled="isLoadingAction"
           data-test="dialog-close-button"
-          @click="showModal = false"
+          @click="modalVisible = false"
         />
       </div>
     </template>
@@ -120,7 +120,7 @@ const props = defineProps({
 const emit = defineEmits(["invalidCode", "invalidToken", "notFound"]);
 
 const isLoadingAction = ref(false);
-const showModal = ref(false);
+const modalVisible = ref(false);
 
 const api = useApi();
 const toast = useToast();